+++ augbuilder/ex_radio.json
@@ -1,68 +1,41 @@
 {
-	"CLAHE": {
-		"clip_limit": [
-			1,
-			4
-		],
-		"tile_grid_size": [
-			8,
-			8
+	"Cutout": {
+		"num_holes": 43,
+		"max_h_size": 100,
+		"max_w_size": 1,
+		"fill_value": [
+			127,
+			80,
+			171
 		],
 		"p": 0.5
 	},
-	"ChannelDropout": {
-		"channel_drop_range": [
-			1,
-			1
+	"ChannelShuffle": {
+		"p": 0.26
+	},
+	"GaussNoise": {
+		"var_limit": [
+			68.79,
+			238.26
 		],
-		"fill_value": 0,
-		"p": 0.5
+		"mean": -75.17,
+		"p": 0.63
 	},
-	"Downscale": {
-		"scale_min": 0.25,
-		"scale_max": 0.25,
-		"interpolation": 0,
-		"p": 0.5
+	"FancyPCA": {
+		"alpha": 16.11
 	},
-	"Equalize": {
-		"mode": "pil",
-		"by_channels": false,
-		"p": 0.1
-	},
-	"GridDistortion": {
-		"num_steps": 10,
-		"distort_limit": [
-			-1.3,
-			1.02
+	"ISONoise": {
+		"color_shift": [
+			0.36,
+			0.68
 		],
-		"interpolation": 4,
-		"border_mode": 4,
-		"value": [
-			255,
-			188,
-			64
+		"intensity": [
+			0.1,
+			1.99
 		],
-		"p": 0.27
+		"p": 0.25
 	},
-	"ImageCompression": {
-		"compression_type": 0,
-		"quality_lower": 80,
-		"quality_upper": 100,
-		"p": 0.5
-	},
-	"RandomRain": {
-		"slant_lower": -15,
-		"slant_upper": 5,
-		"drop_length": 51,
-		"drop_width": 3,
-		"drop_color": [
-			255,
-			44,
-			255
-		],
-		"blur_value": 5,
-		"brightness_coefficient": 0.11,
-		"rain_type": "heavy",
-		"p": 0.18
+	"Flip": {
+		"p": 0.26
 	}
 }
